Real time insights from Twitter
Twitter has this huge influence on the world because of the type of its users. Data from Twitter-storms is available in near real-time. This means we can learn about the big waves of thoughts and moods around the world as they arise.
Here we did a thorough analysis of hot trends